Move util into field_info.rs
