+++ const_typed_builder_derive/src/info/field_info.rs
@@ -1,15 +1,13 @@
-use std::collections::HashSet;
-
+use self::util::{inner_type, is_option};
 use super::struct_info::StructSettings;
-use proc_macro2::Span;
-use quote::format_ident;
-use syn::{ExprPath, Token};
-
 use crate::{
     symbol::{BUILDER, GROUP, MANDATORY, OPTIONAL, PROPAGATE},
-    util::{inner_type, is_option},
     CONST_IDENT_PREFIX,
 };
+use proc_macro2::Span;
+use quote::format_ident;
+use std::collections::HashSet;
+use syn::{ExprPath, Token};
 
 /// Represents the information about a struct field used for code generation.
 #[derive(Debug, PartialEq, Eq)]
@@ -348,5 +346,43 @@ impl FieldSettings {
             }
             Ok(())
         })
+    }
+}
+
+mod util {
+    pub fn is_option(ty: &syn::Type) -> bool {
+        if let syn::Type::Path(type_path) = ty {
+            if type_path.qself.is_some() {
+                return false;
+            }
+            if let Some(segment) = type_path.path.segments.last() {
+                segment.ident == syn::Ident::new("Option", segment.ident.span())
+            } else {
+                false
+            }
+        } else {
+            false
+        }
+    }
+
+    pub fn inner_type(ty: &syn::Type) -> Option<&syn::Type> {
+        let path = if let syn::Type::Path(type_path) = ty {
+            if type_path.qself.is_some() {
+                return None;
+            }
+            &type_path.path
+        } else {
+            return None;
+        };
+        let segment = path.segments.last()?;
+        let syn::PathArguments::AngleBracketed(generic_params) = &segment.arguments else {
+            return None;
+        };
+
+        if let syn::GenericArgument::Type(inner) = generic_params.args.first()? {
+            Some(inner)
+        } else {
+            None
+        }
     }
 }
